+++ packages/eyes-selenium/lib/EyesVisualGrid.js
@@ -1,7 +1,6 @@
 'use strict'
 
 const {makeVisualGridClient, takeDomSnapshot} = require('@applitools/visual-grid-client')
-
 const {
   ArgumentGuard,
   TypeUtils,
@@ -10,8 +9,8 @@ const {
   BrowserType,
   CorsIframeHandle,
   CorsIframeHandler,
-  IgnoreRegionByRectangle,
   VisualGridRunner,
+  showBrowserWarning,
 } = require('@applitools/eyes-sdk-core')
 
 const {Eyes} = require('./Eyes')
@@ -105,6 +104,7 @@ class EyesVisualGrid extends Eyes {
       }
     }
 
+    showBrowserWarning(this._configuration.getBrowsersInfo())
     const {checkWindow, close, abort} = await openEyes(
       this._configuration.toOpenEyesConfiguration(),
     )