+++ packages/eyes-webdriverio-5/src/EyesVisualGrid.js
@@ -12,6 +12,7 @@ const {
   RectangleSize,
   Configuration,
   VisualGridRunner,
+  showBrowserWarning,
 } = require('@applitools/eyes-sdk-core')
 
 const EyesWebDriver = require('./wrappers/EyesWebDriver')
@@ -125,6 +126,7 @@ class EyesVisualGrid extends EyesBase {
       await this.setViewportSize(vs)
     }
 
+    showBrowserWarning(this._configuration.getBrowsersInfo())
     const {checkWindow, close, abort} = await openEyes(
       this._configuration.toOpenEyesConfiguration(),
     )