+++ packages/visual-grid-client/src/sdk/openEyes.js
@@ -122,7 +122,9 @@ function makeOpenEyes({
 
     const supportedBrowsers = getSupportedBrowsers()
     const supportedBrowserKeys = Object.keys(supportedBrowsers)
-    const supportedBrowserKeysStr = `\n* ${supportedBrowserKeys.join('\n* ')}\n`
+    const supportedBrowserKeysStr = `\n* ${supportedBrowserKeys
+      .filter(x => x !== 'edge')
+      .join('\n* ')}\n`
 
     const browsersArray = Array.isArray(browser) ? browser : [browser]
     const browserError = browsersArray.length